feat: add handler when the requested task id is not found

+++ be-todo/src/main/java/com/sesar/betodo/controllers/TaskController.java
@@ -7,7 +7,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Controller
@@ -16,28 +18,52 @@ public class TaskController {
 
     @Autowired
     private TaskService taskService;
+
     @GetMapping("/")
     public ResponseEntity<List<Task>> getAllTasks() {
         return ResponseEntity.ok(taskService.getAllTask());
     }
+
     @GetMapping("/{id}")
     public ResponseEntity<Task> getTaskById(@PathVariable Long id) {
         Optional<Task> task = taskService.getTaskById(id);
         return task.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
+
     @PostMapping("/")
     public ResponseEntity<Task> createTask(@RequestBody Task task) {
         return ResponseEntity.ok(taskService.createNewTask(task));
     }
+
     @PutMapping("/{id}")
-    public ResponseEntity<Task> updateTask(@PathVariable Long id, @RequestBody Task task) {
-        task.setId(id);
-        return ResponseEntity.ok(taskService.updateTask(id, task));
+    public ResponseEntity<Map<String, String>> updateTask(@PathVariable Long id, @RequestBody Task task) {
+        Optional<Task> existingTask = taskService.getTaskById(id);
+        if (existingTask.isPresent()) {
+            task.setId(id);
+            taskService.updateTask(id, task);
+            Map<String, String> response = new HashMap<>();
+            response.put("message", "TODO updated successfully");
+            return ResponseEntity.ok(response);
+        } else {
+            Map<String, String> response = new HashMap<>();
+            response.put("message", "Task not found");
+            return ResponseEntity.status(404).body(response);
+        }
     }
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<Boolean> getAllTasks(@PathVariable Long id) {
-        taskService.deleteTask(id);
-        return ResponseEntity.ok(true);
+    public ResponseEntity<Map<String, String>> deleteTask(@PathVariable Long id) {
+        Optional<Task> existingTask = taskService.getTaskById(id);
+        if (existingTask.isPresent()) {
+            taskService.deleteTask(id);
+            Map<String, String> response = new HashMap<>();
+            response.put("message", "TODO deleted successfully");
+            return ResponseEntity.ok(response);
+        } else {
+            Map<String, String> response = new HashMap<>();
+            response.put("message", "Task not found");
+            return ResponseEntity.status(404).body(response);
+        }
     }
 }
